+++ DataParse/ItemStatistic.cs
@@ -11,9 +11,9 @@ namespace DataParse {
         public float? MeanValue { get; private set; }
         public float? MinValue { get; private set; }
         public float? MaxValue { get; private set; }
-        public double? Cp { get; private set; }
-        public double? Cpk { get; private set; }
-        public double? Sigma { get; private set; }
+        public float? Cp { get; private set; }
+        public float? Cpk { get; private set; }
+        public float? Sigma { get; private set; }
         public int PassCount { get; private set; }
         public int FailCount { get; private set; }
 
@@ -30,21 +30,21 @@ namespace DataParse {
                 MaxValue = listUnNullItems.Max();
 
                 //(sum((Xi-Mean)^2)/count)^0.5
-                Sigma = Math.Sqrt((from r in listUnNullItems
+                Sigma = (float)Math.Sqrt((from r in listUnNullItems
                                    let a = Math.Pow((double)(r - MeanValue), 2)
                                    select a).Sum() / listUnNullItems.Count);
 
-                double? T = null;
-                double? U = null;
-                double? Ca = null;
+                float? T = null;
+                float? U = null;
+                float? Ca = null;
                 if (hl != null && ll != null) {
-                    T = ((double)hl - (double)ll);
-                    U = ((double)hl + (double)ll) / 2;
+                    T = ((float)hl - (float)ll);
+                    U = ((float)hl + (float)ll) / 2;
                     Ca = (MeanValue - U) / (T / 2);
                     //Cp= (Hlimit-Llimit)/(6*Sigma)
-                    Cp = (double)(T / (Sigma * 6));
+                    Cp = (float)(T / (Sigma * 6));
                     //Cpk = Cp*(1-|Ca|)
-                    Cpk = Cp * (1 - Math.Abs((double)Ca));
+                    Cpk = Cp * (1 - Math.Abs((float)Ca));
                 }
             }
